watcher: no need for GetPodInfo in watcher
This patch simplifies the watcher code. It removes the GetPodInfo()
method from the interface, since it does not require any
watcher-specific functionality and can be implemented on its own
instead.
