+++ pkg/process/podinfo_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Tetragon
 
-package watcher
+package process
 
 import (
 	"context"
@@ -10,6 +10,7 @@ import (
 
 	"github.com/cilium/tetragon/api/v1/tetragon"
 	"github.com/cilium/tetragon/pkg/cilium"
+	"github.com/cilium/tetragon/pkg/watcher"
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/wrapperspb"
@@ -43,9 +44,9 @@ func TestK8sWatcher_GetPodInfo(t *testing.T) {
 	assert.NoError(t, err)
 
 	k8sClient := fake.NewSimpleClientset(&pod)
-	watcher := NewK8sWatcher(k8sClient, time.Hour)
+	watcher := watcher.NewK8sWatcher(k8sClient, time.Hour)
 	pid := uint32(1)
-	podInfo, _ := watcher.GetPodInfo("abcd1234", "curl", "cilium.io", 1)
+	podInfo, _ := getPodInfo(watcher, "abcd1234", "curl", "cilium.io", 1)
 	assert.True(t, proto.Equal(podInfo, &tetragon.Pod{
 		Namespace: pod.Namespace,
 		Name:      pod.Name,
