--- pkg/watcher/podinfo.go
@@ -1,24 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of Tetragon
-
-package watcher
-
-import (
-	coreV1 "k8s.io/api/core/v1"
-)
-
-func getExecCommand(probe *coreV1.Probe) []string {
-	if probe != nil && probe.Exec != nil {
-		return probe.Exec.Command
-	}
-	return nil
-}
-
-func getProbes(pod *coreV1.Pod, containerStatus *coreV1.ContainerStatus) ([]string, []string) {
-	for _, container := range pod.Spec.Containers {
-		if container.Name == containerStatus.Name {
-			return getExecCommand(container.LivenessProbe), getExecCommand(container.ReadinessProbe)
-		}
-	}
-	return nil, nil
-}
