+++ pkg/watcher/watcher.go
@@ -10,19 +10,12 @@ import (
 	"strings"
 	"time"
 
-	"github.com/cilium/tetragon/api/v1/tetragon"
-	"github.com/cilium/tetragon/pkg/cilium"
-	"github.com/cilium/tetragon/pkg/filters"
 	"github.com/cilium/tetragon/pkg/logger"
-	"github.com/cilium/tetragon/pkg/metrics/watchermetrics"
-	"google.golang.org/protobuf/types/known/timestamppb"
-	"google.golang.org/protobuf/types/known/wrapperspb"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/informers"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/cache"
 
-	hubblev1 "github.com/cilium/hubble/pkg/api/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -40,9 +33,6 @@ var (
 type K8sResourceWatcher interface {
 	// Find a pod/container pair for the given container ID.
 	FindPod(containerID string) (*corev1.Pod, *corev1.ContainerStatus, bool)
-
-	// Get PodInfo and Endpoint ID for a containerId.
-	GetPodInfo(containerID, binary, args string, nspid uint32) (*tetragon.Pod, *hubblev1.Endpoint)
 }
 
 // K8sWatcher maintains a local cache of k8s resources.
@@ -136,119 +126,6 @@ func (watcher *K8sWatcher) FindPod(containerID string) (*corev1.Pod, *corev1.Con
 		return findContainer(containerID, watcher.podInformer.GetStore().List())
 	}
 	return findContainer(containerID, objs)
-}
-
-func (watcher *K8sWatcher) GetPodInfo(containerID string, binary string, args string, nspid uint32) (*tetragon.Pod, *hubblev1.Endpoint) {
-	if containerID == "" {
-		return nil, nil
-	}
-	pod, container, ok := watcher.FindPod(containerID)
-	if !ok {
-		watchermetrics.GetWatcherErrors("k8s", watchermetrics.FailedToGetPodError).Inc()
-		logger.GetLogger().WithField("container id", containerID).Trace("failed to get pod")
-		return nil, nil
-	}
-	var startTime *timestamppb.Timestamp
-	livenessProbe, readinessProbe := getProbes(pod, container)
-	maybeExecProbe := filters.MaybeExecProbe(binary, args, livenessProbe) ||
-		filters.MaybeExecProbe(binary, args, readinessProbe)
-	if container.State.Running != nil {
-		startTime = timestamppb.New(container.State.Running.StartedAt.Time)
-	}
-
-	ciliumState := cilium.GetCiliumState()
-	endpoint, ok := ciliumState.GetEndpointsHandler().GetEndpointByPodName(pod.Namespace, pod.Name)
-	var labels []string
-	if ok {
-		labels = endpoint.Labels
-	}
-
-	// Don't set container PIDs if it's zero.
-	var containerPID *wrapperspb.UInt32Value
-	if nspid > 0 {
-		containerPID = &wrapperspb.UInt32Value{
-			Value: nspid,
-		}
-	}
-
-	watchermetrics.GetWatcherEvents("k8s").Inc()
-	return &tetragon.Pod{
-		Namespace: pod.Namespace,
-		Name:      pod.Name,
-		Labels:    labels,
-		PodLabels: pod.Labels,
-		Container: &tetragon.Container{
-			Id:   container.ContainerID,
-			Pid:  containerPID,
-			Name: container.Name,
-			Image: &tetragon.Image{
-				Id:   container.ImageID,
-				Name: container.Image,
-			},
-			StartTime:      startTime,
-			MaybeExecProbe: maybeExecProbe,
-		},
-	}, endpoint
-}
-
-// FakeK8sWatcher is used as an "empty" K8sResourceWatcher when --enable-k8s-api flag is not set.
-type FakeK8sWatcher struct {
-	pods []interface{}
-}
-
-// NewK8sWatcher returns a pointer to an initialized FakeK8sWatcher struct.
-func NewFakeK8sWatcher(pods []interface{}) *FakeK8sWatcher {
-	return &FakeK8sWatcher{pods: pods}
-}
-
-// FindPod implements K8sResourceWatcher.FindPod.
-func (watcher *FakeK8sWatcher) FindPod(containerID string) (*corev1.Pod, *corev1.ContainerStatus, bool) {
-	return findContainer(containerID, watcher.pods)
-}
-
-func (watcher *FakeK8sWatcher) GetPodInfo(containerID string, binary string, args string, nspid uint32) (*tetragon.Pod, *hubblev1.Endpoint) {
-	pod, container, ok := watcher.FindPod(containerID)
-	if !ok {
-		watchermetrics.GetWatcherErrors("fake-k8s", watchermetrics.FailedToGetPodError).Inc()
-		logger.GetLogger().WithField("container id", containerID).Trace("failed to get pod")
-		return nil, nil
-	}
-
-	var startTime *timestamppb.Timestamp
-	livenessProbe, readinessProbe := getProbes(pod, container)
-	maybeExecProbe := filters.MaybeExecProbe(binary, args, livenessProbe) ||
-		filters.MaybeExecProbe(binary, args, readinessProbe)
-	if container.State.Running != nil {
-		startTime = timestamppb.New(container.State.Running.StartedAt.Time)
-	}
-
-	var emptyLabels []string
-
-	// Don't set container PIDs if it's zero.
-	var containerPID *wrapperspb.UInt32Value
-	if nspid > 0 {
-		containerPID = &wrapperspb.UInt32Value{
-			Value: nspid,
-		}
-	}
-
-	watchermetrics.GetWatcherEvents("fake-k8s").Inc()
-	return &tetragon.Pod{
-		Namespace: pod.Namespace,
-		Name:      pod.Name,
-		Labels:    emptyLabels,
-		Container: &tetragon.Container{
-			Id:   container.ContainerID,
-			Pid:  containerPID,
-			Name: container.Name,
-			Image: &tetragon.Image{
-				Id:   container.ImageID,
-				Name: container.Image,
-			},
-			StartTime:      startTime,
-			MaybeExecProbe: maybeExecProbe,
-		},
-	}, nil
 }
 
 // TODO(michi) Not the most efficient implementation. Optimize as needed.
